requirements/test.txt
# Test dependencies
-r _base.txt

pytest==3.4.0
pytest-django==3.1.2
